+++ packages/terra-navigation-side-menu/src/constants.js
@@ -1,0 +1,6 @@
+const VARIANTS = {
+  NAVIGATION_SIDE_MENU: 'navigation-side-menu',
+  DRILL_IN: 'drill-in',
+};
+
+export default VARIANTS;