+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added examples for drill-in view under `terra-navigation-side-menu`
+
 ## 1.87.0 - (May 14, 2024)
 
 * Changed